backend/generate/requirements.txt
transformers 
accelerate 
peft 
diffusers
pillow
compel
pika
psycopg2-binary
minio
